Handle GitHub Enterprise subdomain isolation
If a GitHub Enterprise instance has subdomain isolation enabled, it uses a subdomain rather than a path segment to access the codeload service. Because of that we can detect when a request is made for a tarball without `_codeload` in the path, and prefix the hostname with `codeload.`.

I've also updated the regex to allow `/codeload` without an underscore at the beginning. Although I haven't seen this while testing, that's what their current docs say.