Add ability to map hostname for GitHub Enterprise (#12)
If a GitHub Enterprise instance has subdomain isolation enabled, it uses a subdomain rather than a path segment to access the codeload service. Because of that we can detect when a request is made for a tarball without `_codeload` in the path, and prefix the hostname with `codeload.`.

I've also updated the regex to allow `/codeload` without an underscore at the beginning. Although I haven't seen this while testing, that's what their current docs say.

In addition, I've added tests and updated the validation to respect the HTTP method. While doing that I noticed the method was wrong for the "List Installations" request.