Support Azure DO in sub-paths (#28)
Support Azure DevOps instances that are hosted on a child path rather than at the root of the domain. As part of doing this, I've had to merge the rules for listing organization and project repositories because it becomes impossible to distinguish between them with the information we have available.