fix: parsing of the DIAL_INSECURE env var
We were disabling TLS if `SPACELIFT_VCS_AGENT_DIAL_INSECURE` was set to any value, including if it was set to `false`.

While doing this I noticed we had some inconsistency where we were setting most of the agent's fields via the `New` function, but were using an exported field for `HTTPDisableResponseCompression`. I've adjusted it so we use a config object instead to make it obvious what fields are being set.